#108 Create API for search results preview. Renaming address-summary to contract-summary
