--- address-summary/build.gradle
@@ -1,43 +1,0 @@
-project(":address-summary:common") {
-
-    apply plugin: "io.spring.dependency-management"
-    jar.archiveName = "common-address-summary"
-
-    dependencies {
-
-        compile project(":common")
-        compile project(":common-kafka")
-        compile project(":cassandra-service")
-
-        compile 'org.springframework.boot:spring-boot-starter'
-
-        compile 'org.springframework.boot:spring-boot-starter-webflux'
-        compile 'org.springframework.boot:spring-boot-starter-reactor-netty'
-        compile 'org.springframework.boot:spring-boot-starter-actuator'
-        compile 'io.micrometer:micrometer-registry-prometheus'
-        compile 'io.micrometer:micrometer-core'
-        compile 'org.jetbrains.kotlinx:kotlinx-coroutines-jdk8'
-    }
-}
-
-project(":address-summary:bitcoin") {
-
-    apply plugin: "org.springframework.boot"
-
-    dependencies {
-
-        compile project(":address-summary:common")
-    }
-}
-
-
-project(":address-summary:ethereum") {
-
-    apply plugin: "org.springframework.boot"
-
-    dependencies {
-
-        compile project(":address-summary:common")
-
-    }
-}
