--- address-summary/common/src/main/kotlin/fund/cyber/address/common/CommonConfiguration.kt
@@ -1,22 +1,0 @@
-package fund.cyber.address.common
-
-import fund.cyber.search.model.chains.Chain
-import io.micrometer.core.instrument.MeterRegistry
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.actuate.autoconfigure.metrics.MeterRegistryCustomizer
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.EnableScheduling
-
-@EnableScheduling
-@Configuration
-class CommonConfiguration {
-
-    @Autowired
-    private lateinit var chain: Chain
-
-    @Bean
-    fun metricsCommonTags(): MeterRegistryCustomizer<MeterRegistry> {
-        return MeterRegistryCustomizer { registry -> registry.config().commonTags("chain", chain.name) }
-    }
-}
