--- address-summary/common/src/main/kotlin/fund/cyber/address/common/delta/apply/AddressLockException.kt
@@ -1,3 +1,0 @@
-package fund.cyber.address.common.delta.apply
-
-class AddressLockException: Exception()
