+++ contract-summary/bitcoin/Dockerfile
