+++ contract-summary/build.gradle
@@ -1,0 +1,43 @@
+project(":contract-summary:common") {
+
+    apply plugin: "io.spring.dependency-management"
+    jar.archiveName = "common-contract-summary"
+
+    dependencies {
+
+        compile project(":common")
+        compile project(":common-kafka")
+        compile project(":cassandra-service")
+
+        compile 'org.springframework.boot:spring-boot-starter'
+
+        compile 'org.springframework.boot:spring-boot-starter-webflux'
+        compile 'org.springframework.boot:spring-boot-starter-reactor-netty'
+        compile 'org.springframework.boot:spring-boot-starter-actuator'
+        compile 'io.micrometer:micrometer-registry-prometheus'
+        compile 'io.micrometer:micrometer-core'
+        compile 'org.jetbrains.kotlinx:kotlinx-coroutines-jdk8'
+    }
+}
+
+project(":contract-summary:bitcoin") {
+
+    apply plugin: "org.springframework.boot"
+
+    dependencies {
+
+        compile project(":contract-summary:common")
+    }
+}
+
+
+project(":contract-summary:ethereum") {
+
+    apply plugin: "org.springframework.boot"
+
+    dependencies {
+
+        compile project(":contract-summary:common")
+
+    }
+}
