+++ contract-summary/common/src/main/kotlin/fund/cyber/contract/common/delta/apply/ContractLockException.kt
@@ -1,0 +1,3 @@
+package fund.cyber.contract.common.delta.apply
+
+class ContractLockException: Exception()
