+++ contract-summary/common/src/main/resources/application.yml
