+++ contract-summary/common/src/main/resources/banner.txt
